LayoutTests/imported/w3c/web-platform-tests/html/anonymous-iframe/anonymous-iframe-popup.tentative.https.window-expected.txt


FAIL Cross-origin popup from normal/credentiallessiframes. assert_equals: Opener from credentialless iframe should be blocked. expected null but got object "[object Window]"
FAIL Same-origin popup from normal/credentialless iframes. assert_equals: Opener from credentialless iframe should be blocked. expected null but got object "[object Window]"
